Refactored code, fixed bugs and enable unit tests
