+++ src/main/java/ImageHoster/controller/CommentController.java
@@ -1,0 +1,37 @@
+package ImageHoster.controller;
+
+import ImageHoster.model.Image;
+import ImageHoster.model.User;
+import ImageHoster.service.CommentService;
+import ImageHoster.service.ImageService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RequestParam;
+
+import javax.servlet.http.HttpSession;
+
+@Controller
+public class CommentController {
+
+    @Autowired
+    private CommentService commentService;
+
+    @Autowired
+    private ImageService imageService;
+
+    @RequestMapping(value = "/image/{imageId}/{imageTitle}/comments", method = RequestMethod.POST)
+    public String addComment(@PathVariable Integer imageId, @PathVariable String imageTitle, @RequestParam("comment") String commentText, Model model, HttpSession session) {
+        commentService.addComment(commentText, imageService.getImage(imageId), (User) session.getAttribute("loggeduser"));
+
+        Image image = imageService.getImage(imageId);
+        model.addAttribute("image", image);
+        model.addAttribute("tags", image.getTags());
+        model.addAttribute("comments", image.getComments());
+        return "redirect:/images/" + image.getId() + "/" + image.getTitle();
+    }
+
+}
