+++ src/main/java/ImageHoster/repository/ImageRepository.java
@@ -111,17 +111,4 @@ public class ImageRepository {
         }
     }
 
-    public void addComment(Comment comment) {
-        EntityManager em = emf.createEntityManager();
-        EntityTransaction transaction = em.getTransaction();
-
-        try {
-            transaction.begin();
-            em.persist(comment);
-            transaction.commit();
-        } catch (Exception e) {
-            transaction.rollback();
-        }
-    }
-
 }
